LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-lanes/tentative/track-sizing/auto-repeat/intrinsic-auto-repeat/column-auto-repeat-auto-016-expected.txt


FAIL Auto-fit repeat tracks with auto size and auto/explicit item placement assert_equals: expected "200px 200px 200px 200px 0px 200px 0px 0px 0px 0px" but got "166.65625px 166.65625px 166.65625px 166.65625px 166.65625px 166.65625px"
